Django_Backend/appointment_Management/requirements.txt
agora-token-builder==1.0.0
asgiref==3.7.2
Babel==2.13.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.1
cryptography==41.0.5
Django==4.2.6
django-appointment==2.1.1
django-cors-headers==4.3.0
django-otp==1.2.4
django-phonenumber-field==7.2.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
ndg-httpsclient==0.5.1
phonenumbers==8.13.25
phonenumberslite==8.13.23
Pillow==10.1.0
pusher==3.3.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
pyotp==2.9.0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7